Refactor price assignment in product creation for consistency
- Updated the price assignment in the generateProductFromImage method to ensure it is cast to a float, improving data consistency for product pricing.

+++ Modules/Product/Services/AiProductService.php
@@ -46,7 +46,7 @@ class AiProductService
         // Create a default product variant
         $product->variants()->create([
             'name' => $productDetails['name'],
-            'price' => $productDetails['suggested_price'] ?? 0,
+            'price' => (float) ($productDetails['suggested_price'] ?? 0.00),
             'price_currency' => 'SAR',
             'type' => 'count',
             'stock' => 0,
